+++ branches/RC_0_15/src/error_code.cpp
@@ -212,7 +212,7 @@ namespace libtorrent
 			"",
 			"",
 		};
-		if (ev < 0 || ev >= sizeof(msgs)/sizeof(msgs[0]))
+		if (ev < 0 || ev >= int(sizeof(msgs)/sizeof(msgs[0])))
 			return "Unknown error";
 		return msgs[ev];
 	}
